# demo/demo.openapi.yaml
openapi: 3.0.0
info:
  title: Express Inspectr API
  version: 1.0.0
  description: A sample API for testing purposes
paths:
...
  /api/services/inspectr:
    get:
      summary: Get service info
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ServiceInfo'
    post:
      summary: Post to inspectr service
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PingRequest'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ServiceInfo'
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    put:
      summary: Update service
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ServiceInfo'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ServiceInfo'
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    delete:
      summary: Delete service
      responses:
        '204':
          description: No Content
...
components:
  schemas:
    ServiceInfo:
      type: object
      properties:
        name:
          type: string
          example: "Inspectr Service"
        version:
          type: string
          example: "1.0.1"
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
          example: "Missing required parameters"
    PingRequest:
      type: object
      properties:
        message:
          description: The ping message
          type: string
          example: "Welcome the Express Inspectr"
        user:
          description: The user sending the ping
          type: string
          example: "Marco Polo"
      required:
        - message
        - user